Add test to check if the room already exist
